+++ cmd/hidden_lake/applications/messenger/internal/handler/chat_ws.go
@@ -1,22 +1,22 @@
 package handler
 
 import (
-	"github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/internal/receiver"
+	"github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/internal/msgbroker"
+	"github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/internal/utils"
 	"golang.org/x/net/websocket"
 )
 
-func FriendsChatWS(pReceiver receiver.IMessageReceiver) func(pWS *websocket.Conn) {
+func FriendsChatWS(pBroker msgbroker.IMessageBroker) func(pWS *websocket.Conn) {
 	return func(pWS *websocket.Conn) {
 		defer pWS.Close()
 
-		subscribe := new(receiver.SMessage)
+		subscribe := new(utils.SSubscribe)
 		if err := websocket.JSON.Receive(pWS, subscribe); err != nil {
 			return
 		}
 
-		pReceiver.Init()
 		for {
-			msg, ok := pReceiver.Recv(subscribe.FAddress)
+			msg, ok := pBroker.Consume(subscribe.FAddress)
 			if !ok {
 				return
 			}