+++ cmd/hidden_lake/applications/messenger/internal/utils/types.go
@@ -1,8 +1,12 @@
 package utils
 
-type SMessageInfo struct {
+type SSubscribe struct {
+	FAddress string `json:"address"`
+}
+
+type SMessage struct {
 	FFileName  string `json:"filename"` // can be ""
 	FTimestamp string `json:"timestamp"`
 	FPseudonym string `json:"pseudonym"`
-	FMessage   string `json:"message"`
+	FMainData  string `json:"maindata"`
 }